chore: refactor utils folder by grouping and reorganizing
