+++ src/index.ts
@@ -1,4 +1,3 @@
 export { nextTag } from './tag';
-export { releaseNotes } from './notes';
 export { lastTagForPrefix, filterCommits } from './git';
 export * from './types';
