--- src/utils/getDateFromCommit.ts
@@ -1,3 +1,0 @@
-export const getDateFromCommit = (dateWithTime: string): string => {
-  return /(\d{4}-\d{2}-\d{2})/.exec(dateWithTime)?.[0] ?? '';
-};
