+++ src/utils/os.test.ts
@@ -1,4 +1,4 @@
-import { execCmd } from './execCmd';
+import { execCmd } from './os';
 
 describe('when executing cmd', () => {
   it('should run multiple lines', async () => {
